Refactor details and compare.
