+++ src/SecurityCompare/index.ts
@@ -33,7 +33,6 @@ import SecurityCompareConnector from './SecurityCompareConnector';
 
 
 export * from './SecurityCompareConnector';
-export * from '../SecurityDetails/SecurityDetailsConverter';
 export * from '../SecurityDetails/SecurityDetailsOptions';
