+++ src/SecurityDetails/SecurityDetailsJSON.ts
@@ -128,12 +128,6 @@ namespace SecurityDetailsJSON {
         );
     }
 
-    export function isSecurityCompareResponse (
-        json: Array<SecurityDetailsResponse>
-    ): boolean { 
-        return json.length > 1;
-    }
-
 
     function isTrailingPerformance (
         json?: unknown
