+++ src/SecurityDetails/index.ts
@@ -34,8 +34,7 @@ import SecurityDetailsConnector from './SecurityDetailsConnector';
 
 export * from './SecurityDetailsConnector';
 export * from './SecurityDetailsOptions';
-export * from './SecurityDetailsConverter';
-export * from './Converters';
+export * from '../Shared/Converters';
 
 
 /* *
