+++ tests/SecurityDetails/SecurityDetailsConnector.test.ts
@@ -18,9 +18,8 @@ export async function securityDetailsLoad (
     );
 
     Assert.ok(
-        connector.converter instanceof
-        MC.SecurityDetailsConverter,
-        'Converter should be instance of SecurityDetailsConverter.'
+        connector.converter instanceof MC.TrailingPerformanceConverter,
+        'Converter should be instance of TrailingPerformanceConverter.'
     );
 
     await connector.load();
@@ -141,8 +140,18 @@ export async function countryExposureLoad (
     await connector.load();
 
     Assert.deepStrictEqual(
-        connector.table.getColumnNames()[0],
-        'CountryExposure_Assets',
+        connector.table.getColumnNames(),
+        [
+            'CountryExposure_Assets',
+            'CountryExposure_NotClassified',
+            'CountryExposure_Type',
+            'CountryExposure_Bond_L',
+            'CountryExposure_Bond_S',
+            'CountryExposure_Bond_N',
+            'CountryExposure_Equity_L',
+            'CountryExposure_Equity_S',
+            'CountryExposure_Equity_N'
+        ],
         'Country exposure table should exist of expected columns.'
     );
 }
